+++ data/index.ts
@@ -37,7 +37,9 @@ export * from './cube/Cube';
 export * from './cube/CubeField';
 export * from './cube/Query';
 export * from './cube/View';
+export * from './cube/ViewRow';
 export * from './cube/ViewRowData';
+export * from './cube/row/RowUpdate';
 
 export * from './validation/constraints';
 export * from './validation/Rule';